refatorando codigo

+++ parking-project/src/main/java/br/com/fcamara/parkingproject/controller/dto/ParkingLotDto.java
@@ -6,8 +6,8 @@ import lombok.Getter;
 @Getter
 public class ParkingLotDto {
     private Long id;
-    private Integer motocycle_spaces;
-    private Integer car_spaces;
+    private Long motocycle_spaces;
+    private Long car_spaces;
 
     public ParkingLotDto(ParkingLot parkingLot) {
         this.id = parkingLot.getId();
